# fix(html/parser): perf and fixes