--- a/crates/swc_html_parser/tests/fixture/text/lf/dom.rust-debug
+++ b/crates/swc_html_parser/tests/fixture/text/lf/dom.rust-debug
@@ -67,3 +67,25 @@
 }
 
 "
+|     "
+"
+|     <div>
+|       test="test
+
+test
+
+test"
+|     "
+"
+|     <div>
+|       test="
+
+test
+
+test
+
+test
+
+"
+|     "
+"