--- a/crates/swc_html_parser/tests/fixture/text/lf/input.html
+++ b/crates/swc_html_parser/tests/fixture/text/lf/input.html
@@ -53,3 +53,17 @@ test
 }
 
 </style>
+<div test="test
+
+test
+
+test"></div>
+<div test="
+
+test
+
+test
+
+test
+
+"></div>