… crates/swc_html_parser/tests/fixture/text/cr/input.html
… crates/swc_html_parser/tests/fixture/text/cr/output.json
… crates/swc_html_parser/tests/fixture/text/cr/span.rust-debug
… crates/swc_html_parser/tests/fixture/text/lf/output.json
… crates/swc_html_parser/tests/fixture/text/lf/span.rust-debug
… crates/swc_html_parser/tests/recovery/attribute/duplicate/output.stderr
… crates/swc_html_parser/tests/recovery/attribute/missing-value/output.json