<!-- Note: CI script will automatically rebase your PR so please do not rebase unless required -->

<!-- Thanks for submitting a pull request! Please provide enough information so that others can review your pull request. -->

**Description:**

Here:
- perf improvement - avoid new string allocation and use only two buffers
- fix broken span in recovery mode
- fix bug with raw and newlines in attribute values
- simplify code - avoid using `LexResult` where it is unnecessary
- update tests

Perf:
```
Benchmarking html/parser_document/css_2021_spec: Collecting 100 samples in estimated 5.0453 s (200 itera                                                                                                        html/parser_document/css_2021_spec                        
                        time:   [24.893 ms 24.993 ms 25.102 ms]
                        change: [-9.3727% -8.5518% -7.7744%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 9 outliers among 100 measurements (9.00%)
  6 (6.00%) high mild
  3 (3.00%) high severe

Benchmarking html/parser_document/github_com_17_05_2022: Collecting 100 samples in estimated 5.6538 s (6                                                                                                        html/parser_document/github_com_17_05_2022                        
                        time:   [9.3805 ms 9.4121 ms 9.4473 ms]
                        change: [-12.224% -10.727% -9.2354%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 6 outliers among 100 measurements (6.00%)
  4 (4.00%) high mild
  2 (2.00%) high severe

Benchmarking html/parser_document/stackoverflow_com_17_05_2022: Collecting 100 samples in estimated 5.03                                                                                                        html/parser_document/stackoverflow_com_17_05_2022                        
                        time:   [8.3459 ms 8.4242 ms 8.5158 ms]
                        change: [-12.482% -10.492% -8.5345%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 9 outliers among 100 measurements (9.00%)
  6 (6.00%) high mild
  3 (3.00%) high severe

Benchmarking html/parser_document_fragment/css_2021_spec: Collecting 100 samples in estimated 5.0795 s (                                                                                                        html/parser_document_fragment/css_2021_spec                        
                        time:   [24.988 ms 25.257 ms 25.612 ms]
                        change: [-7.4636% -6.3417% -5.0886%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 14 outliers among 100 measurements (14.00%)
  6 (6.00%) high mild
  8 (8.00%) high severe

Benchmarking html/parser_document_fragment/github_com_17_05_2022: Collecting 100 samples in estimated 5.                                                                                                        html/parser_document_fragment/github_com_17_05_2022                        
                        time:   [9.4571 ms 9.5606 ms 9.6936 ms]
                        change: [-3.2604% -2.2218% -0.6482%] (p = 0.00 < 0.05)
                        Change within noise threshold.
Found 15 outliers among 100 measurements (15.00%)
  7 (7.00%) high mild
  8 (8.00%) high severe

Benchmarking html/parser_document_fragment/stackoverflow_com_17_05_2022: Collecting 100 samples in estim                                                                                                        html/parser_document_fragment/stackoverflow_com_17_05_2022                        
                        time:   [8.4049 ms 8.4661 ms 8.5385 ms]
                        change: [-5.4380% -4.3263% -3.2211%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 6 outliers among 100 measurements (6.00%)
  4 (4.00%) high mild
  2 (2.00%) high severe

```

**BREAKING CHANGE:**

Yes, if you will use Lexer directly

**Related issue (if exists):**

No