@alexander-akait 2022-11-30
@kdy1 Ready for review, there are many changes, but most of them just rewriting `String::new()`/`String::with_capacity(number)` on buffers